requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
anyio==4.13.0
asgiref==3.11.1
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==6.0.4
django-axes==8.3.1
django-cors-headers==4.9.0
django-environ==0.13.0
django-filter==25.2
django-htmx==1.27.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
filelock==3.25.2
fsspec==2026.3.0
gunicorn==25.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.10.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kombu==5.6.2
lxml==6.0.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
packaging==26.0
pgvector==0.4.2
pillow==12.2.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
Pygments==2.20.0
PyJWT==2.12.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
PyYAML==6.0.3
redis==7.4.0
regex==2026.4.4
rich==14.3.4
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.5.3
typer==0.24.1
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
vine==5.1.0
watchdog==6.0.0
wcwidth==0.6.0
